fix: types for providers
