+++ packages/ui-kit/src/components/AccessTokenProvider/AccessTokenProvider.tsx
@@ -1,28 +1,7 @@
 import React, { createContext, useCallback, useEffect, useRef, useState } from 'react'
-import { AccessTokenError } from './utils'
 import { sleep } from '../../helpers'
 import { useLog } from '../Log'
-
-export interface AccessTokenContextValue {
-  /**
-   * The current access token, if any.
-   *
-   * If the AccessTokenProvider was instantiated without a token, or the fetch token function has not completed, this property will be undefined.
-   *
-   * Otherwise, when re-fetching an access token, this property will remain equal to the previous access token until the fetch token function completes again.
-   */
-  accessToken?: string | null
-  /**
-   * If true, the AccessTokenProvider is fetching an access token.
-   *
-   * The AccessTokenProvider could be fetching an initial access token, or it could be re-fetching the access token.
-   */
-  isLoading?: boolean
-  /**
-   * If present, the AccessTokenProvider encountered an error fetching the access token. This will be set to `undefined` as soon as fetching the access token succeeds.
-   */
-  error?: Error
-}
+import { AccessTokenContextValue, AccessTokenProviderProps } from './AccessTokenProvider.types'
 
 const ACCESS_TOKEN_REFRESH_INTERVAL = 3300000 // 55 minutes
 const ACCESS_TOKEN_MAX_RETRIES = 3
@@ -30,20 +9,11 @@ const ACCESS_TOKEN_RETRY_INTERVAL = 1000
 
 export const AccessTokenContext = createContext<AccessTokenContextValue>({})
 
-export interface AccessTokenProviderProps {
-  /**
-   * Function that the provider will use to fetch the access token.
-   * @returns {Promise<string>} A promise that resolves to the access token
-   */
-  fetchToken?: () => Promise<string>
-  /**
-   * If passed, the provider will ignore the `fetchToken` function and pass this access token to all the children components.
-   */
-  accessToken?: string
-  children?: React.ReactNode
-}
-
-export const AccessTokenProvider: React.FC<AccessTokenProviderProps> = ({ children, accessToken, fetchToken }) => {
+export const AccessTokenProvider: React.FC<React.PropsWithChildren<AccessTokenProviderProps>> = ({
+  children,
+  accessToken,
+  fetchToken
+}) => {
   const [isLoading, setIsLoading] = useState(accessToken == null)
   const [fetchedToken, setFetchedToken] = useState<string | undefined | null>(undefined)
   const [error, setError] = useState<Error | undefined>(undefined)
@@ -71,13 +41,15 @@ export const AccessTokenProvider: React.FC<AccessTokenProviderProps> = ({ childr
         setFetchedToken(token)
 
         break
-      } catch (error: any) {
+      } catch (error) {
+        const fetchError = error as Error
+
         if (!mounted.current) break
         log.warn('Failed to fetch access token', error)
 
         if (retries === ACCESS_TOKEN_MAX_RETRIES) {
           setIsLoading(false)
-          setError(new AccessTokenError('Failed to fetch access token: ' + error.message))
+          setError(new Error('Failed to fetch access token: ' + fetchError.message))
         }
 
         retries++
