+++ packages/ui-kit/src/components/FilterProvider/FilterProvider.types.ts
@@ -1,0 +1,12 @@
+import { FilterInput } from '../../helpers'
+
+export interface FilterInputWithId extends FilterInput {
+  id: symbol
+}
+
+export interface FilterContextValue {
+  /** Filters that will be provided to the child components */
+  filters: FilterInputWithId[]
+  /** Setter function for the Filters */
+  setFilters: React.Dispatch<React.SetStateAction<FilterInputWithId[]>>
+}
